move e2e utils to dedicated pkg for future use
