+++ pkg/utils/operatorutil/operatorutil.go
