+++ pkg/utils/operatorutil/syntheticutil.go
@@ -19,8 +19,8 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/GoogleCloudPlatform/prometheus-engine/e2e/kubeutil"
 	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/operator"
+	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/utils/kubeutil"
 	appsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
